# api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Plexus Data API",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/sources": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "List Sources",
        "operationId": "list_sources_v1_sources_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "online",
                    "offline"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/sources/{source_id}": {
      "get": {
        "tags": [
          "sources"
        ],
        "summary": "Get Source",
        "operationId": "get_source_v1_sources__source_id__get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/sources/{source_id}/logs": {
      "get": {
        "tags": [
          "logs"
        ],
        "summary": "Get Logs",
        "operationId": "get_logs_v1_sources__source_id__logs_get",
        "parameters": [
          {
            "name": "source_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Source Id"
            }
          },
          {
            "name": "tail",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10000
                },
                {
                  "type": "null"
                }
              ],
              "description": "last N events, newest-first then reversed to chronological",
              "title": "Tail"
            },
            "description": "last N events, newest-first then reversed to chronological"
          },
          {
            "name": "last",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "e.g. 1h, 30m, 7d",
              "title": "Last"
            },
            "description": "e.g. 1h, 30m, 7d"
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601 start time",
              "title": "Start"
            },
            "description": "ISO 8601 start time"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO 8601 end time",
              "title": "End"
            },
            "description": "ISO 8601 end time"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "default": 1000,
              "title": "Limit"
            }
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "filter by event name",
              "title": "Name"
            },
            "description": "filter by event name"
          },
          {
            "name": "x-api-key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "X-Api-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LogEvent"
                  },
                  "title": "Response Get Logs V1 Sources  Source Id  Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceListResponse": {
        "properties": {
          "devices": {
            "items": {
              "$ref": "#/components/schemas/DeviceResponse"
            },
            "type": "array",
            "title": "Devices"
          }
        },
        "type": "object",
        "required": [
          "devices"
        ],
        "title": "DeviceListResponse"
      },
      "DeviceResponse": {
        "properties": {
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "online": {
            "type": "boolean",
            "title": "Online"
          },
          "last_seen_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen Ms"
          }
        },
        "type": "object",
        "required": [
          "source_id",
          "online",
          "last_seen_ms"
        ],
        "title": "DeviceResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LogEvent": {
        "properties": {
          "timestamp_ms": {
            "type": "integer",
            "title": "Timestamp Ms"
          },
          "metric": {
            "type": "string",
            "title": "Metric"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "tags": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Tags",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "timestamp_ms",
          "metric",
          "value"
        ],
        "title": "LogEvent"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}